+++ administrator/com_blank/tmpl/panel/default.php
